+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- New Member Profile layout including contact phone and email
 
 ## [2.1.7] - 2020-10-14
 ## Fixed
@@ -28,7 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Improves Member Profile formatting
-- Re-organizes organizes Profile and Account settings and related menus items
+- Re-organizes Profile and Account settings and related menus items
 
 ### Fixed
 - Returns to sign-up wizard on page reload if incomplete